backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
playwright==1.40.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
google-generativeai==0.3.2
pydantic==2.5.0
APScheduler==3.10.4
fake-useragent==1.4.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
httpx==0.25.2
pandas==2.1.4
